pyproject.toml
[tool.poetry]
name = "construction-repairs"
version = "0.1.0"
description = ""
authors = ["kirill_barkhatov <k.s.barkhatov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
line_length = 119
